Refactor BundleService and remove MapService dependency
Refactored the `BundleService` class to remove the dependency on `MapService` by updating its constructor to only take `ILogger<TagService>` and `IOptions<PathSettings>` as parameters. Removed the logic that used `MapService` to get extra bundle information and update the `ShopConfig`. Additionally, removed the `ClaimToGuid` dictionary and its related logic from the `MapService` class, as this mapping is no longer required.