Refactor Profiles class and update GetProfiles method
Changed method functionality to accept user IDs in the request body, returning a `BadRequest` if none are provided.
Implemented processing of user IDs to retrieve profiles for "switch" and "uplay" platforms, returning an `Ok` result with a list of `PlatformProfile` objects.
Introduced a new `PlatformProfile` class to encapsulate platform-specific profile data.